// skill-stack-frontend/src/app/models/shared.enums.ts
export enum ProficiencyLevel {
  Beginner = 'Beginner',
  Novice = 'Novice',
  Intermediate = 'Intermediate',
  Advanced = 'Advanced',
  Expert = 'Expert',
}

export enum Status {
  Completed = 'Completed',
  InProgress = 'In Progress',
  NotStarted = 'Not Started',
}

export enum Priority {
  Low = 'Low',
  Medium = 'Medium',
  High = 'High',
}

export enum TimeLine {
  ShortTerm = 'Short-Term',
  MedTerm = 'Med-Term',
  LongTerm = 'Long-Term',
}